fix: 不要な文などを削除
