+++ package.json
@@ -18,7 +18,6 @@
     "@mui/material": "^5.4.3",
     "@types/react-syntax-highlighter": "^13.5.2",
     "dayjs": "^1.10.8",
-    "diff-in-time": "^1.0.3",
     "firebase": "^8.10.1",
     "next": "12.1.0",
     "react": "17.0.2",
